CF/requirements.txt
twilio
google-cloud-logging
google-cloud-secret-manager